requirements.txt
numpy >= 1.11
matplotlib>=1.5.1
scipy>=0.9
biopython >= 1.69
pandas>=0.18
pysam>= 0.9
nose >= 1.3.7
mock >= 2.0.0